Cleanup the README situation
There is no need to install README.debian or README.known_issues.
Besides seeming out-of-date, they detail issues (mostly build issues)
which are not relevant to the packaged version of imapproxy.

We then patch README to remove references to those files and build
instructions as well as update paths and executable names to match those
used in this package.